Refactored Assembly resolver for clarity and improved logging & commenting. Also added check for VC12 runtime, which outputs a warning if no sign of C++ Redistributable.
Changed Prequisities validation to raise exception by default if VC runtime appears to be missing. This can be suppressed (treated as a warning) by appsetting in cases where the validation may be imperfect

Handle the 'standard case' better. We were not handling the case where the Engine DL was in base directory, because of the suffix to the name. Rather than add complexity to the resolver, change the Interop to request the correct platform assembly and let .NET do it's standard probing.