Random unhandled exceptions creating wire geometry
Multi-threaded usage of OCC function BRepOffsetAPI_MakeOffset.Perform()
leads to an Unpredictable behavior and unhandled native access violation
exceptions. That is because function BRepOffsetAPI_MakeOffset.Perform()
uses static BRepMAT2d_Explorer Exp, so it should not be used in a
multi-threaded mode.
As a solution I have wrapped this call in a thread lock.