More tolerant geometry creation of models
Fixes: #534

We had an issue for files (e.g. Autodesk Civil 3D) that use a composite
 curve instead of a gradient curve when an alignment is not mapped
to a surface.

The fix is done via a workaround that injects a GradientCurve placed
with a flat verticalAlignment, because compositeCurve
is not implemented in the 2d curve generation. Appropriate tests are
added.

Curvefactory is changed for ifcGradientCurve to prevent multiple attempt
of generating complex geometries when they are known to fail.
This is achieved keeping a log of them
via HashSet<int>^ attemptedEntityLabels.

CurveFactory::BuildCurve2d(IIfcCurve^ curve, XCurveType% curveType)
was improved to deal with the switch label that did not
appropriately manage XCurveType::IfcGradientCurve.

ModelPlacementBuilder was fixed when dealing with linearPlacements
(a clerical error referenced the wrong variable in the
RelativePlacement).

XbimPlacementTree has been fixed to fail more gracefully when the
geometry engine fails to compute the matrix for a linearPlacement.

GetOpeningsAndProjections() was fixed to prevent unrecoverable errors
on null RelatingObjects.