Geometry parallelism bug and minor fixes
1) TopoDS_Wire WireFactory::BuildWire(IIfcIndexedPolyCurve^, bool) threw
an exception in opencascade when run in release outside of the
development environment for some models.

This only happens when multi-threading, so I've added a std::lock_guard
to prevent aggressive parallelism of that specific call.

2) added logging tracing functions to FactoryBase.h

3) Added options to the regression executable to set the required
engine version (v5 or v6). Also the /caching arg accepts an extra arg
to differentiate the extension of the retained xbim cache files.